requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
tiktoken
pypdf
